Parsing OK

+++ libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: chsiffre <chsiffre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/04 14:08:10 by charles           #+#    #+#             */
-/*   Updated: 2022/11/24 16:38:27 by chsiffre         ###   ########.fr       */
+/*   Updated: 2023/01/17 13:22:54 by chsiffre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,7 +27,7 @@ typedef struct s_list
 size_t			ft_strlen(const char *s);
 size_t			ft_strlcpy(char *dst, const char *src, size_t dstsize);
 size_t			ft_strlcat(char *dst, const char *src, size_t dstsize);
-int				ft_atoi(const char *str);
+long long		ft_atoi(const char *str);
 int				ft_strncmp(const char *s1, const char *s2, size_t n);
 int				ft_isalnum(int c);
 int				ft_isascii(int c);
